Read details (VERY IMPORTANT)
So, now we have a new store "patient.js". With this store, we can get the: UUID, First Name, Last Name, Date of Birth, Social (Security Number).
AT THE MOMENT OF THIS PUSH: ONLY THE UUID IS SET, THE OTHER VALUES WILL BE EQUAL TO NULL!!
To be added: email, phone?

Another store like this should be created for the "trusted person".

The request to create Observations is working and have been tested, feel free to try it out as you wish.